refactor: consolidate LinusBrainAreaPopup into ActivityDetectionPopup
AreaView showed two separate chips whenever Brain was installed: Activity
Detection (always shown, localized, with history graphs and the sensor
list) and a second "Linus Brain" chip opening LinusBrainAreaPopup — a
near-duplicate covering the same activity/presence/duration/stats ground in
a different flat-cards layout with hardcoded English strings (no
Helper.localize at all). Two popups answering the same question two
different ways depending on installed integrations, rather than one model
consistently augmented when Brain is available.

Everything LinusBrainAreaPopup showed beyond what ActivityDetectionPopup
already covers was: automatic lighting switch and a device config link —
both genuinely Brain-only, no other home in the UI. Folded those into a new
Controls section (Brain-gated) in ActivityDetectionPopup; dropped the
redundant all-lights control (the area's own light card already provides
it) and the "Time Occupied" stat (already shown in the existing Statistics
section, same entity). Removed the LinusBrainAreaPopup chip/import from
AreaView and deleted the now-dead file.
