+++ custom_components/linus_dashboard/translations/en.json
@@ -138,7 +138,11 @@
           "add_sensor_explanation": "Click to assign sensors to this area",
           "add_entity_button": "Add Entity",
           "manage_area_entities": "Manage area entities",
-          "auto_detection_criteria": "ℹ️ **Auto-detection**\n\nEntities appear here if they are assigned to this area AND:\n✅ Binary sensors with `device_class` = motion, occupancy, or presence\n✅ Media players\n\n**To add sensors:**\n1. Click the button below to open entity settings\n2. Assign your sensors to this area\n3. Configure the `device_class` if needed\n4. **Click the refresh badge (🔄) at the top** to update the list\n\n[📖 Learn more about Activity Detection](https://github.com/Thank-you-Linus/Linus-Dashboard/blob/main/docs/ACTIVITY_DETECTION.md)"
+          "auto_detection_criteria": "ℹ️ **Auto-detection**\n\nEntities appear here if they are assigned to this area AND:\n✅ Binary sensors with `device_class` = motion, occupancy, or presence\n✅ Media players\n\n**To add sensors:**\n1. Click the button below to open entity settings\n2. Assign your sensors to this area\n3. Configure the `device_class` if needed\n4. **Click the refresh badge (🔄) at the top** to update the list\n\n[📖 Learn more about Activity Detection](https://github.com/Thank-you-Linus/Linus-Dashboard/blob/main/docs/ACTIVITY_DETECTION.md)",
+          "controls_title": "Controls",
+          "controls_subtitle": "Area automation",
+          "automatic_lighting": "Automatic Lighting",
+          "view_device_config": "View Device Configuration"
         }
       },
       "lights_group_popup": {
